Expose ArrayDefinitionRepository
