+++ src/main/java/io/luna/game/model/def/DefinitionRepository.java
@@ -28,42 +28,6 @@ import static com.google.common.base.Preconditions.checkState;
 public abstract class DefinitionRepository<T extends Definition> implements Iterable<T> {
 
     /**
-     * An implementation of {@link DefinitionRepository} that is backed internally by an array.
-     */
-    static final class ArrayDefinitionRepository<T extends Definition> extends DefinitionRepository<T> {
-
-        /**
-         * An array of definitions.
-         */
-        public final T[] definitions;
-
-        /**
-         * Creates a new {@link ArrayDefinitionRepository}.
-         *
-         * @param length The length of the backing array.
-         */
-        @SuppressWarnings("unchecked")
-        public ArrayDefinitionRepository(int length) {
-            definitions = (T[]) new Definition[length];
-        }
-
-        @Override
-        void put(int id, T definition) {
-            definitions[id] = definition;
-        }
-
-        @Override
-        public Optional<T> get(int id) {
-            return Optional.ofNullable(definitions[id]);
-        }
-
-        @Override
-        public Iterator<T> newIterator() {
-            return Iterators.forArray(definitions);
-        }
-    }
-
-    /**
      * An implementation of {@link DefinitionRepository} that is backed internally by a {@link LinkedHashMap}.
      */
     static final class MapDefinitionRepository<V extends Definition> extends DefinitionRepository<V> {
